+++ models/br_me_comex_stat/br_me_comex_stat__ncm_exportacao.sql
@@ -29,9 +29,9 @@ with
             ) sigla_uf_ncm,
             safe_cast(id_via as string) id_via,
             safe_cast(id_urf as string) id_urf,
-            safe_cast(quantidade_estatistica as int64) quantidade_estatistica,
-            safe_cast(peso_liquido_kg as int64) peso_liquido_kg,
-            safe_cast(valor_fob_dolar as int64) valor_fob_dolar
+            safe_cast(quantidade_estatistica as float64) quantidade_estatistica,
+            safe_cast(peso_liquido_kg as float64) peso_liquido_kg,
+            safe_cast(valor_fob_dolar as float64) valor_fob_dolar
         from {{ set_datalake_project("br_me_comex_stat_staging.ncm_exportacao") }} as t
         {% if is_incremental() %}
             where
